gestion_tickets_app/requirements.txt
amqp==5.2.0
asgiref==3.8.1
astroid==3.1.0
billiard==4.2.0
black==24.4.2
blinker==1.7.0
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
dill==0.3.8
Django==5.0.4
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.3
kombu==5.3.7
Markdown==3.6
MarkupSafe==2.1.5
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.1
pluggy==1.5.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pylint==3.1.0
pytest==8.2.0
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
tomlkit==0.12.4
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.2